fix: separate orderbook components and set default view in orderbook chart
