+++ .lintstagedrc.json
@@ -1,3 +1,3 @@
 {
-  "*.js|jsx|ts|tsx": ["npm run --prefix frontend lint:fix"]
+  "*.@(js|jsx|ts|tsx|json)": ["npm run --prefix frontend lint:fix"]
 }
