+++ frontend/src/app/coin/[market]/ClientPage.tsx
@@ -2,14 +2,14 @@
 
 import { useEffect, useState } from 'react';
 import { useParams } from 'next/navigation';
-import OrderbookList from '@/app/coin/components/OrderbookList';
-import CandleChart from '@/app/coin/components/CandleChart';
-import TradeList from '@/app/coin/components/TradeList';
-import NewsList from '@/app/coin/components/NewsList';
 import { useWebSocket } from '@/context/WebSocketContext';
 import type { CandleItem, CandleChartDto } from '@/types';
 import axios from 'axios';
 import { generateMockNews } from '@/lib/utils';
+import OrderbookList from '../components/orderbook/OrderbookList';
+import CandleChart from '../components/CandleChart';
+import TradeList from '../components/TradeList';
+import NewsList from '../components/NewsList';
 
 export default function ClientPage() {
   const { market } = useParams() as { market: string };
