+++ frontend/src/app/coin/components/orderbook/index.ts
@@ -1,0 +1,1 @@
+export * from './OrderbookList';
